requirements.txt
openpyxl==3.1.*
pyyaml==6.0.*
jinja2==3.1.*
jsonschema==4.22.*
click==8.1.*
rich==13.*
